feat: more more cleanup
